azure-pdf-search/requirements.txt
azure-functions
azure-core
azure-search-documents
openai
